Write function that sums two arguments
This function is meant to sum two arguments as well as every number
in between both arguments. It returns an error message if the
argument is a negative or non-numeric value. It works if the smaller
number is passed as the first argument as well as the other way around.