Write function to determine if leap year
This function takes one argument that is called into a few boolean
comparison statements. If the argument is not at all divisible by 4,
then it immediately returns false. If the argument is divisible by 4,
divisible by 100, but not divisible by 400, returns false because that
is not a leap year. All arguments that are divisible by 4 and 400 will
return true.